Consolidate non-null/missing property tests to parameterized suite
